% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw04pjiarong0829-package.R
\docType{package}
\name{hw04pjiarong0829-package}
\alias{hw04pjiarong0829}
\alias{hw04pjiarong0829-package}
\title{hw04pjiarong0829: Functions is for STAT-613 homework 4}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Jiarong Li \email{jl0606a@american.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
